+++ landingzones/tests/test_views_ajax.py
@@ -4,8 +4,10 @@ import json
 import random
 import string
 
+from django.test import override_settings
 from django.urls import reverse
 
+from landingzones.constants import ZONE_STATUS_MOVED
 from landingzones.tests.test_views import ViewTestBase
 from landingzones.views_ajax import STATUS_TRUNCATE_LEN
 
@@ -19,14 +21,16 @@ class TestZoneStatusRetrieveAjaxView(ViewTestBase):
             'landingzones:ajax_status',
             kwargs={'project': self.project.sodar_uuid},
         )
+        self.post_data = {
+            'zones': {str(self.zone.sodar_uuid): {'modified': ''}}
+        }
         self.post_kw = {'content_type': 'application/json'}
 
     def test_post(self):
         """Test ZoneStatusRetrieveAjaxView POST"""
-        post_data = {'zones': {str(self.zone.sodar_uuid): {'modified': ''}}}
         with self.login(self.user):
             response = self.client.post(
-                self.url, data=json.dumps(post_data), **self.post_kw
+                self.url, data=json.dumps(self.post_data), **self.post_kw
             )
         self.assertEqual(response.status_code, 200)
         expected = {
@@ -37,7 +41,8 @@ class TestZoneStatusRetrieveAjaxView(ViewTestBase):
                 'truncated': False,
             }
         }
-        self.assertEqual(response.data, expected)
+        self.assertEqual(response.data['zones'], expected)
+        self.assertEqual(response.data['zone_create_limit'], False)
 
     def test_post_modified(self):
         """Test ZoneStatusRetrieveAjaxView POST"""
@@ -53,7 +58,9 @@ class TestZoneStatusRetrieveAjaxView(ViewTestBase):
                 self.url, data=json.dumps(post_data), **self.post_kw
             )
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {})
+        self.assertEqual(
+            response.data, {'zones': {}, 'zone_create_limit': False}
+        )
 
     def test_post_truncate(self):
         """Test POST with truncated status_info"""
@@ -62,10 +69,9 @@ class TestZoneStatusRetrieveAjaxView(ViewTestBase):
             for _ in range(STATUS_TRUNCATE_LEN * 2)
         )
         self.zone.save()
-        post_data = {'zones': {str(self.zone.sodar_uuid): {'modified': ''}}}
         with self.login(self.user):
             response = self.client.post(
-                self.url, data=json.dumps(post_data), **self.post_kw
+                self.url, data=json.dumps(self.post_data), **self.post_kw
             )
         self.assertEqual(response.status_code, 200)
         expected = {
@@ -76,7 +82,8 @@ class TestZoneStatusRetrieveAjaxView(ViewTestBase):
                 'truncated': True,
             }
         }
-        self.assertEqual(response.data, expected)
+        self.assertEqual(response.data['zones'], expected)
+        self.assertEqual(response.data['zone_create_limit'], False)
 
     def test_post_no_zone(self):
         """Test POST with no zones"""
@@ -86,7 +93,30 @@ class TestZoneStatusRetrieveAjaxView(ViewTestBase):
                 self.url, data=json.dumps(post_data), **self.post_kw
             )
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {})
+        self.assertEqual(
+            response.data, {'zones': {}, 'zone_create_limit': False}
+        )
+
+    @override_settings(LANDINGZONES_ZONE_CREATE_LIMIT=1)
+    def test_post_limit(self):
+        """Test POST with zone creation limit reached"""
+        with self.login(self.user):
+            response = self.client.post(
+                self.url, data=json.dumps(self.post_data), **self.post_kw
+            )
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.data['zone_create_limit'], True)
+
+    @override_settings(LANDINGZONES_ZONE_CREATE_LIMIT=1)
+    def test_post_limit_existing_finished(self):
+        """Test POST with zone creation limit and finished zone"""
+        self.zone.set_status(ZONE_STATUS_MOVED)
+        with self.login(self.user):
+            response = self.client.post(
+                self.url, data=json.dumps(self.post_data), **self.post_kw
+            )
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.data['zone_create_limit'], False)
 
 
 class TestZoneStatusInfoRetrieveAjaxView(ViewTestBase):